refactor(errors) rename files, validators
