+++ src/middlewares/errors/errors-middleware.ts
@@ -1,12 +1,10 @@
 import { Errback, ErrorRequestHandler, NextFunction, Request, Response } from "express";
-import { CustomError } from "../errors/errors";
-import { STATUS_CODES } from "../errors/status-codes";
+import { CustomError } from "./custom-errors";
+import { STATUS_CODES } from "./status-codes";
 import mongoose, { Error } from "mongoose";
 
 export const errHandleMiddleware = (err: CustomError, req: Request, res: Response, next: NextFunction) => {
   let { statusCode, message } = err;
-
-  console.log(err instanceof mongoose.Error.ValidationError);
 
   if (err instanceof mongoose.Error.ValidationError) {
     statusCode = STATUS_CODES.BAD_REQUEST.statusCode;
